Fix windows build (again)
