+++ checks/container.go
@@ -139,30 +139,6 @@ func fmtContainers(
 	return containers
 }
 
-// chunkContainers chunks the ctrList into a slice of chunks using a specific
-// number of chunks. len(result) MUST EQUAL chunks.
-func chunkedContainers(
-	ctrList []*model.Container,
-	chunks int) [][]*model.Container {
-	perChunk := (len(ctrList) / chunks) + 1
-	chunked := make([][]*model.Container, chunks)
-	chunk := make([]*model.Container, 0, perChunk)
-	i := 0
-	for _, ctr := range ctrList {
-		chunk = append(chunk, ctr)
-
-		if len(chunk) == perChunk {
-			chunked[i] = chunk
-			chunk = make([]*model.Container, 0, perChunk)
-			i++
-		}
-	}
-	if len(chunk) > 0 {
-		chunked[i] = chunk
-	}
-	return chunked
-}
-
 func calculateCtrPct(cur, prev, sys2, sys1 uint64, numCPU int, before time.Time) float32 {
 	now := time.Now()
 	diff := now.Unix() - before.Unix()
